+++ ouds_core/lib/components/radio_button/ouds_radio_button.dart
@@ -11,6 +11,9 @@
  * //
  */
 
+/// OudsRadioButton
+library;
+
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter_svg/flutter_svg.dart';