+++ ouds_core/lib/components/tag/internal/ouds_tag_control_state.dart
@@ -9,6 +9,9 @@
 // Software description: Flutter library of reusable graphical components
 //
 
+/// @nodoc
+library;
+
 /// Enum representing the state of the tag control.
 enum OudsTagControlState {
   enabled,